refactor: aside layer 트리구조 변경 후 컴포넌트 분리 및 커스텀 훅 제작
