+++ front/packages/user/src/components/aside/layer/LayerTilesetContent.tsx
@@ -1,0 +1,420 @@
+import React, {createRef, RefObject, useEffect, useRef, useState} from 'react';
+import {DndProvider} from "react-dnd";
+import {getBackendOptions, MultiBackend, NodeModel, Tree, TreeMethods} from "@minoru/react-dnd-treeview";
+import {
+    CreateUserGroupInput, LayerAssetType,
+    LayersetDeleteAssetDocument, Maybe,
+    Mutation, Query, RemoteDocument, RemoteQueryVariables,
+    UserLayerAsset, UserLayerGroup
+} from "@mnd/shared/src/types/layerset/gql/graphql.ts";
+import {useGlobeController} from "@/components/providers/GlobeControllerProvider.tsx";
+import {useMutation} from "@tanstack/react-query";
+import {layersetGraphqlFetcher} from "@/api/queryClient.ts";
+import {RESTORE_USERLAYER, SAVE_USERLAYER} from "@/graphql/layerset/Mutation.ts";
+import {useMutation as apolloUseMutation} from "@apollo/client/react/hooks/useMutation";
+import {useRecoilState, useSetRecoilState} from "recoil";
+import {
+    InitialOpenState,
+    layersState,
+    NodeModelsState,
+    UserLayerGroupState,
+    visibleToggledLayerIdsState,
+    visibleToggledLayerIdState
+} from "@/recoils/Layer.ts";
+import {useLazyQuery} from "@apollo/client";
+import {produce} from "immer";
+import * as Cesium from "cesium";
+import {GET_USERLAYERGROUPS} from "@/graphql/layerset/Query.ts";
+import {
+    getLayersFromNodeModels,
+    layerGroupsToNodemodels,
+    nodeModelIdToNumber, nodeModlesToCreateUserGroupInput
+} from "@/components/aside/layer/LayerNodeModel.ts";
+
+const LayerTilesetContent = () => {
+    const {initialized, globeController} = useGlobeController();
+    const {mutateAsync: saveUserLayerMutateAsync} = useMutation({
+        mutationFn:({input}:{input:CreateUserGroupInput[]}) => layersetGraphqlFetcher<Mutation>(SAVE_USERLAYER, {input})
+    });
+    const {mutateAsync: restoreUserLayerMutateAsync} = useMutation({
+        mutationFn:() => layersetGraphqlFetcher<Mutation>(RESTORE_USERLAYER)
+    });
+    const [deleteAssetMutation] = apolloUseMutation(LayersetDeleteAssetDocument);
+
+    const [treeData, setTreeData] = useRecoilState<NodeModel[]>(NodeModelsState);
+    const a = useRef<RefObject<HTMLDivElement>[]>([]);
+    const [userLayerGroups, setUserLayerGroups] = useRecoilState<Maybe<UserLayerGroup>[]>(UserLayerGroupState);
+    const setLayers = useSetRecoilState<UserLayerAsset[]>(layersState);
+    const setVisibleToggledLayerId = useSetRecoilState<string | null>(visibleToggledLayerIdState);
+    const setVisibleToggledLayerIds = useSetRecoilState<{ids:string[], visible:boolean} | null>(visibleToggledLayerIdsState);
+    const [visibleAll, setVisibleAll] = useState<boolean>(false);
+    const [initialOpen, setinitialOpen] = useRecoilState<number[]>(InitialOpenState);
+    const treeRef = useRef<TreeMethods>(null);
+    const [getRemoteData, { data: remoteData }] = useLazyQuery(RemoteDocument);
+    const [remoteType, setRemoteType] = useState('');
+   const [originalTreeData, setOriginalTreeData] = useState<NodeModel[]>(treeData);
+
+
+    const setNodeModels = (groups:Maybe<UserLayerGroup>[]): NodeModel[] => {
+        const initOpen: number[] = [];
+
+        const nodeModels = layerGroupsToNodemodels(groups);
+        setTreeData(nodeModels);
+        setOriginalTreeData(nodeModels);
+
+        nodeModels.filter((nodeModel:NodeModel) => nodeModel.parent === 0).forEach((nodeModel:NodeModel) => {
+            if (!(nodeModel.data as UserLayerGroup).collapsed) {
+                initOpen.push(nodeModelIdToNumber(nodeModel.id));
+            }
+        });
+
+        setinitialOpen(initOpen);
+        return nodeModels;
+    }
+
+    const setLayersFromNodeModels = (nodeModels:NodeModel[]) => {
+        const assets = getLayersFromNodeModels(nodeModels);
+        setLayers(produce((draft) => {
+            draft.length = 0;
+            draft.push(...assets);
+        }));
+    }
+
+// // 검색 필터링 함수 수정
+//     const searchFilterNodeModels = (nodeModels: NodeModel[], searchTerm: string): NodeModel[] => {
+//         if (!searchTerm) return nodeModels;
+//
+//         const lowerCaseSearchTerm = searchTerm.toLowerCase().replace(/\s+/g, '');
+//         console.log(lowerCaseSearchTerm);
+//
+//         const layerGroups = nodeModels.filter(node => node.parent === 0);
+//         const layers = nodeModels.filter(node => node.parent !== 0);
+//
+//         const filteredLayers = layers.filter(nodeModel => {
+//             const nodeName = nodeModel.text.toLowerCase().replace(/\s+/g, '');
+//             console.log(nodeName);
+//             return nodeName.includes(lowerCaseSearchTerm);
+//         });
+//
+//         console.log(filteredLayers);
+//
+//         const parentGroupIds = new Set(filteredLayers.map(layer => layer.parent));
+//
+//         const filteredNodeModels = layerGroups
+//             .filter(group => parentGroupIds.has(group.id))
+//             .concat(filteredLayers);
+//
+//         return filteredNodeModels;
+//     };
+//
+//     console.log(originalTreeData)
+// // useEffect 수정
+//     useEffect(() => {
+//         if (searchTerm) {
+//             // 검색 시 원본 데이터를 기준으로 필터링
+//             const filteredNodeModels = searchFilterNodeModels(originalTreeData, searchTerm);
+//             setTreeData(filteredNodeModels);
+//             setinitialOpen(filteredNodeModels.map(node => nodeModelIdToNumber(node.id)));
+//         } else {
+//             setTreeData(originalTreeData);
+//         }
+//     }, [searchTerm, originalTreeData]);
+
+    useEffect(() => {
+        if (remoteData) {
+            if (!initialized) return;
+            const viewer = globeController?.viewer;
+            if (!viewer) return;
+
+            if (remoteType === LayerAssetType.Raster) {
+                const { latLonBoundingBox } = remoteData.remote.coverage;
+                viewer.camera.flyTo({
+                    destination: Cesium.Rectangle.fromDegrees(latLonBoundingBox.minx, latLonBoundingBox.miny, latLonBoundingBox.maxx, latLonBoundingBox.maxy),
+                    duration: 2,
+                });
+            } else if (remoteType === LayerAssetType.Vector) {
+                const { latLonBoundingBox } = remoteData.remote.featureType;
+                viewer.camera.flyTo({
+                    destination: Cesium.Rectangle.fromDegrees(latLonBoundingBox.minx, latLonBoundingBox.miny, latLonBoundingBox.maxx, latLonBoundingBox.maxy),
+                    duration: 2,
+                });
+            }
+        }
+    }, [remoteData, initialized, remoteType]);
+
+    useEffect(() => {
+        if (userLayerGroups.length === 0) {
+            layersetGraphqlFetcher<Query>(GET_USERLAYERGROUPS)
+                .then((result) => {
+                    const {userGroups} = result;
+                    setUserLayerGroups(userGroups);
+                    const nodeModels = setNodeModels(userGroups);
+                    setLayersFromNodeModels(nodeModels);
+                });
+        }
+    }, [userLayerGroups]);
+
+    if (userLayerGroups.length === 0) return null;
+
+    const handleDrop = (newTree: NodeModel[]) => {
+        setTreeData(newTree);
+        setOriginalTreeData(newTree);
+        setLayersFromNodeModels(newTree);
+    };
+
+    const toggleCollapse = (node:NodeModel, isOpen:boolean) => {
+        const id = node.id;
+        const tree = treeRef.current;
+
+        if (!tree) return;
+        let collapse = false;
+        if (isOpen) {
+            collapse = true;
+        }
+
+        const assetIdx = treeData.findIndex((node:NodeModel) => node.id === id);
+        setTreeData(produce((draft) => {
+            (draft[assetIdx].data as UserLayerGroup).collapsed = collapse;
+        }));
+        setOriginalTreeData(produce((draft) => {
+            (draft[assetIdx].data as UserLayerGroup).collapsed = collapse;
+        }));
+        setinitialOpen(produce((draft) => {
+            const a = nodeModelIdToNumber(id);
+            if (collapse) {
+                draft.splice(draft.indexOf(a), 1);
+            } else {
+                draft.push(a);
+            }
+        }));
+    }
+    const toggleVisibleEventHandle = (asset:NodeModel) => {
+        const assetIdx = treeData.findIndex((node:NodeModel) => node.id === asset.id);
+        setTreeData(produce((draft) => {
+            const visible = (draft[assetIdx].data as UserLayerAsset).visible;
+            (draft[assetIdx].data as UserLayerAsset).visible = !visible;
+        }));
+        setOriginalTreeData(produce((draft) => {
+            const visible = (draft[assetIdx].data as UserLayerAsset).visible;
+            (draft[assetIdx].data as UserLayerAsset).visible = !visible;
+        }));
+
+        setVisibleToggledLayerId((asset.data as UserLayerAsset).assetId.toString());
+    }
+
+    const toggleLayer = () => {
+        const layerIds:string[] = [];
+        setTreeData(produce((draft) => {
+            draft.forEach((nodeModel:NodeModel) => {
+                if (nodeModel.parent !== 0) {
+                    const asset = nodeModel.data as UserLayerAsset;
+                    layerIds.push(asset.assetId.toString());
+                    asset.visible = visibleAll;
+                }
+            });
+        }));
+        setOriginalTreeData(produce((draft) => {
+            draft.forEach((nodeModel:NodeModel) => {
+                if (nodeModel.parent !== 0) {
+                    const asset = nodeModel.data as UserLayerAsset;
+                    layerIds.push(asset.assetId.toString());
+                    asset.visible = visibleAll;
+                }
+            });
+        }));
+        setVisibleToggledLayerIds({ids:layerIds, visible:visibleAll});
+        setVisibleAll(!visibleAll);
+    };
+
+    const restoreToDefault = async () => {
+        if(!confirm('시스템 설정으로 복원하시겠습니까?')) return;
+
+        const result = await restoreUserLayerMutateAsync();
+        const {saveUserLayer } = result;
+
+        const nodeModels = setNodeModels(saveUserLayer);
+        setLayersFromNodeModels(nodeModels);
+    };
+
+    const saveState = async () => {
+        if(!confirm('현 설정을 저장하시겠습니까?')) return;
+
+        const input = nodeModlesToCreateUserGroupInput(treeData);
+        await saveUserLayerMutateAsync({input});
+    };
+
+    function extractPath(url: string): string {
+        const parsedUrl = new URL(url);
+        return parsedUrl.pathname + parsedUrl.search + parsedUrl.hash;
+    }
+
+    const flyToLayer = (node:NodeModel) => {
+        const asset = node.data;
+        const layerType = asset?.type;
+        const viewer = globeController?.viewer;
+        if (!viewer) return;
+        setRemoteType(layerType);
+
+        switch (layerType) {
+            case LayerAssetType.Tiles3D: {
+                const model = Cesium.Cesium3DTileset.fromUrl(asset?.properties.resource);
+                viewer.flyTo(model, {duration: 2});
+                return;
+            }
+            case LayerAssetType.Raster:
+            case LayerAssetType.Vector: {
+                const { properties } = asset;
+                const { layer } = properties;
+                const { resource } = layer;
+
+                let variables: RemoteQueryVariables = { href : resource.href };
+
+                if (import.meta.env.MODE === 'production') {
+                    const path = extractPath(resource.href);
+
+                    const { protocol, hostname, port } = window.location;
+
+                    const portPart = port ? `:${port}` : '';
+
+                    const href = `${protocol}//${hostname}${portPart}${path}`;
+                    variables =  { href };
+                }
+                getRemoteData({
+                    variables: variables
+                });
+                return
+            }
+            case LayerAssetType.Layergroup: {
+                const {properties} = asset;
+                const {layerGroup} = properties;
+                const {minx, maxx, miny, maxy} = layerGroup.bounds;
+                const extent = Cesium.Rectangle.fromDegrees(minx, miny, maxx, maxy);
+                viewer.camera.flyTo({
+                    destination: extent,
+                    duration: 2,
+                });
+                return;
+            }
+            case LayerAssetType.Cog:
+            default: {
+                alert("허용되지 않는 형태입니다.");
+                return;
+            }
+
+        }
+
+    }
+
+    const deleteLayer = (node: NodeModel) => {
+        if (!confirm('레이어를 삭제하시겠습니까?')) return;
+
+        try {
+            return deleteAssetMutation({
+                variables: { ids: node.data.assetId }
+            })
+                .then(() => {
+                    alert('레이어가 성공적으로 삭제되었습니다.');
+                    setTreeData((prevTreeData) => {
+                        const newTreeData = prevTreeData.filter((treeNode) => treeNode.id !== node.id);
+                        return newTreeData;
+                    });
+                    setOriginalTreeData((prevTreeData) => {
+                        const newTreeData = prevTreeData.filter((treeNode) => treeNode.id !== node.id);
+                        return newTreeData;
+                    });
+                    setLayersFromNodeModels((prevTreeData) => prevTreeData.filter((treeNode) => treeNode.id !== node.id));
+                });
+        } catch (error) {
+            console.error(error);
+            alert('레이어 삭제 중 오류가 발생했습니다. 다시 시도해주세요.');
+        }
+    };
+
+    return (
+        <div className="layer-scroll">
+            {/*<li className="button-area">*/}
+            {/*    <div className="layer-button">*/}
+            {/*        <button type="button" onClick={toggleLayer}*/}
+            {/*                className={`layer-funtion-button ${!visibleAll ? 'visible' : 'not-visible'}`}></button>*/}
+            {/*        <button type="button" onClick={restoreToDefault} className="layer-funtion-button">R</button>*/}
+            {/*        <button type="button" onClick={saveState} className="layer-funtion-button">S</button>*/}
+            {/*    </div>*/}
+            {/*</li>*/}
+            <DndProvider backend={MultiBackend} options={getBackendOptions()}>
+                <Tree
+                    ref={treeRef}
+                    tree={treeData}
+                    rootId={0}
+                    classes={{container: 'layer-list',}}
+                    initialOpen={initialOpen}
+                    sort={false}
+                    canDrop={(tree, {dragSource, dropTargetId, dropTarget}) => {
+                        if (!dragSource) return;
+
+                        if (
+                            (dragSource.parent === 0 && dropTarget === undefined)
+                            || (dragSource?.parent === dropTargetId)
+                            || (dragSource.parent && !dropTarget?.parent)
+                        ) {
+                            return true;
+                        } else if (
+                            (dragSource.parent === 0 && dropTarget?.parent === 0)
+                        ) {
+                            return false;
+                        }
+                    }}
+                    render={(node, {depth, isOpen, onToggle, hasChild}) => {
+                        const ref = createRef<HTMLDivElement>();
+                        a.current.push(ref);
+                        return (
+                            <>
+                                {
+                                    node.parent === 0 ? (
+                                            <div className="group-node-container" onClick={() => {
+                                                toggleCollapse(node, isOpen)
+                                            }}>
+                                                <div className="group-node">
+                                                    <div className="group-icon">G</div>
+                                                    <div className="group-name">{node.text}</div>
+                                                </div>
+                                            </div>
+                                        )
+                                        : (
+                                            <div className="layer-node-container" key={node.id}>
+                                                <span className="type type-3d"></span>
+                                                <span className="layer-name">{node.text}</span>
+                                                <div className="layer-button">
+                                                    <button type="button"
+                                                            onClick={() => {
+                                                                toggleVisibleEventHandle(node)
+                                                            }}
+                                                            className={`layer-funtion-button ${(node.data as UserLayerAsset).visible ? 'visible' : 'not-visible'}`}></button>
+                                                    {/*<label className="switch">*/}
+                                                    {/*    <input type="checkbox" onChange={toggleVisible} checked={(node.data as LayerAsset).visible ?? false}/>*/}
+                                                    {/*    <span className="slider"></span>*/}
+                                                    {/*</label>*/}
+                                                    <button type="button"
+                                                            onClick={() => {
+                                                                flyToLayer(node)
+                                                            }}
+                                                            className="layer-funtion-button map-view"></button>
+                                                    <button type="button"
+                                                            onClick={() => {
+                                                                deleteLayer(node)
+                                                            }}
+                                                            className="layer-funtion-button delete"></button>
+                                                </div>
+                                            </div>
+                                        )
+                                }
+                            </>
+                        )
+                    }}
+                    onDrop={handleDrop}
+                />
+            </DndProvider>
+        </div>
+    );
+};
+
+export default LayerTilesetContent;
